pyproject.toml
[project]
name = "flywise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu==1.7.4",
    "fastapi==0.109.0",
    "groq==1.0.0",
    "httpx==0.28.1",
    "nltk==3.8.1",
    "numpy==1.26.3",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "scikit-learn==1.4.0",
    "sentence-transformers==2.3.1",
    "spacy>=3.7",
    "torch==2.1.2",
    "transformers==4.36.2",
    "uvicorn[standard]==0.27.0",
]
